% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdt.data.R
\name{sdt.data}
\alias{sdt.data}
\title{Create data for one sample}
\usage{
sdt.data(tot.signal = 85, tot.lure = 100, min.hits = 0.5, min.cr = 0.6)
}
\arguments{
\item{tot.signal}{A number - The total number of signals i.e. blue dots that have to be detected Default = 85}

\item{tot.lure}{A number - The total number of lures i.e. number green dots that should not be responded to, Default = 100}

\item{min.hits}{A number - Minimum ammount of hits. Dafalt to 0.5}

\item{min.cr}{A number - Minimum ammount of correct rejections.  Dafalt to 0.6}
}
\description{
This creates a dataset with 1000 participants to be used with the main function \code{\link{rate.statistics}}.
}
\examples{
\dontrun{
sdt.data(tot.signal = 85, tot.lure = 100, min.hits = 0.5, min.cr = 0.6)
}
}
