% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdt.data2.R
\name{sdt.data2}
\alias{sdt.data2}
\title{Create data for 2 samples/groups}
\usage{
sdt.data2(tot.signal = 85, tot.lure = 100, min.hits = 0.7, min.cr = 0.8,
  min.hits2 = 0.3, min.cr2 = 0.3)
}
\arguments{
\item{tot.signal}{A number - The total number of signals i.e. blue dots that have to be detected Default = 85}

\item{tot.lure}{A number - The total number of lures i.e. number green dots that should not be responded to, Default = 100}

\item{min.hits}{A number - Minimum ammount of hits for group 1. Dafalt to 0.7}

\item{min.cr}{A number - Minimum ammount of correct rejections for group 1.  Dafalt to 0.8}

\item{min.hits2}{A number - Minimum ammount of hits for group 2. Dafalt to 0.3}

\item{min.cr2}{A number - Minimum ammount of correct rejections for group 2.  Dafalt to 0.3}
}
\description{
This creates a dataset with 1000 participants to be used with the main function.
\code{\link{rate.statistics2}}
}
\examples{
\dontrun{
sdt.data2(tot.signal = 85, tot.lure = 100,
          min.hits = 0.7, min.cr = 0.8,
          min.hits2 = 0.3, min.cr2 = 0.3)
}
}
