% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdt.data.R
\name{sdt.data}
\alias{sdt.data}
\title{Create data for one sample}
\usage{
sdt.data(tot.signal = 85, tot.lure = 100, min.hits = 0.5, min.cr = 0.6)
}
\arguments{
\item{tot.signal}{A number - The total number of signals i.e. blue dots that have to be detected Default = 85}

\item{tot.lure}{A number - The total number of lures i.e. number green dots that should not be responded to, Default = 100}

\item{min.hits}{A number - Minimum proportion of hits. Dafalt to 0.5}

\item{min.cr}{A number - Minimum proportion of correct rejections.  Dafalt to 0.6}
}
\value{
A data table with the following columns
   \item{hits}{total number of hits per participant randomly chosen between tot.signal and the minimum proportion hits specified}
   \item{corRej}{total number of correct rejection per participant randomly chosen between tot.lure and the minimum proportion critical rejections specified}
   \item{miss}{total number of misses per participant - left from the total signals after removing hits}
   \item{falarm}{total number of false alarms per participant - left from the total lures after removing correct rejections}
}
\description{
Creates a dataset with 1000 participants to be used with the main function \code{\link{rate.statistics}}.
The parameters can be changed or it can be ran with the defaults by calling sdt.data()
}
\examples{
   sdt.data()

   sdt.data(tot.signal = 15, tot.lure = 150, min.hits = 0.8, min.cr = 0.5)

}
