% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdt.data2.R
\name{sdt.data2}
\alias{sdt.data2}
\title{Create data for 2 samples/groups}
\usage{
sdt.data2(tot.signal = 85, tot.lure = 100, min.hits = 0.7, min.cr = 0.8,
  min.hits2 = 0.3, min.cr2 = 0.3)
}
\arguments{
\item{tot.signal}{A number - The total number of signals i.e. blue dots that have to be detected Default = 85}

\item{tot.lure}{A number - The total number of lures i.e. number green dots that should not be responded to, Default = 100}

\item{min.hits}{A number - Minimum ammount of hits for group 1. Dafalt to 0.7}

\item{min.cr}{A number - Minimum ammount of correct rejections for group 1.  Dafalt to 0.8}

\item{min.hits2}{A number - Minimum ammount of hits for group 2. Dafalt to 0.3}

\item{min.cr2}{A number - Minimum ammount of correct rejections for group 2.  Dafalt to 0.3}
}
\value{
A data table with the following columns
   \item{hits}{total number of hits per participant randomly chosen between tot.signal and the minimum proportion hits specified}
   \item{corRej}{total number of correct rejection per participant randomly chosen between tot.lure and the minimum proportion critical rejections specified}
   \item{miss}{total number of misses per participant - left from the total signals after removing hits}
   \item{falarm}{total number of false alarms per participant - left from the total lures after removing correct rejections}
   \item{hits2}{total number of hits per participant for group 2 randomly chosen between tot.signal and the minimum proportion hits specified}
   \item{corRej2}{total number of correct rejection per participant for group 2 randomly chosen between tot.lure and the minimum proportion critical rejections specified}
   \item{miss2}{total number of misses per participant for group 2- left from the total signals after removing hits}
   \item{falarm2}{total number of false alarms per participant for group 2 - left from the total lures after removing correct rejections}
}
\description{
This creates a dataset with 1000 participants to be used with the main function.
\code{\link{rate.statistics2}}. It can be run with its defaults sdt.rmcs() or by specifying parameters.
}
\examples{
    sdt.data2()
    sdt.data2(tot.signal = 100, tot.lure = 50,
          min.hits = 0.45, min.cr = 0.3,
          min.hits2 = 0.2, min.cr2 = 0.9)
}
